+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
     "cSpell.ignoreWords": [
         "edsger",
-        "herokuapp"
+        "herokuapp",
+        "powerstats"
     ]
 }
