+++ reactreduxapp/src/App.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import './App.css';
-import MySuperHero from './components/MySuperHero';
+import SuperHero from './store/MySuperHero';
 
 function App() {
   return (
     <div className="App">
       <h1>Superhero Info: </h1>
-      <MySuperHero />
+      <SuperHero />
     </div>
   );
 }
