+++ reactreduxapp/src/store/Reducers/Superheros.js
