+++ reactreduxapp/src/store/actions/Superhero.js
@@ -10,23 +10,10 @@ export function getSuperHero() {
         dispatch({FETCHING_SUPERHERO_START});
         axios.get('https://superheroapi.com/api/10221162178461894/search/1')
         .then( res => {
-            dispatch({type:FETCHING_SUPERHERO_SUCCESS, payload: res.data})
+            dispatch({type:FETCHING_SUPERHERO_SUCCESS, payload: res.data.results})
         })
         .catch(err => {
             dispatch({type:FETCHING_SUPERHERO_FAILURE, payload:err})
         })
     }
-
-    // dispatch({ type: FETCHING_QUOTE_START });
-
-    // axios.get('https://programming-quotes-api.herokuapp.com/quotes')
-    //     .then(res => dispatch({ type:FETCHING_QUOTE_SUCCESS, payload: res.data.results}))
-    //     .catch(err => {
-    //         console.log("err", err);
-
-    //     dispatch({
-    //         type: FETCHING_QUOTE_FAILURE,
-    //         payload: err
-    //     });
-    // });
 }
